A few updates to the code and I've started Q6
I had to ensure the necessary class attributes would be integer values, I added a test class for each question and then moved it to be beneath the code for that specific question. I also created a function and test class that will check to see that the pokedex itself is not empty.